[feature] Add support for non-JSON-ifyable properties
Until now properties that are not stringifyable as JSON
and reparsable cannot be used. This not only includes complex
objects but also: undefined, NaN.

Currently the only complex data type supported is Date.
But via the provided plugininterface it is very easy to
add additional types (also from code that is using this library)